moved out test pre-setup to a separate object
